requirements.txt
aioredis==1.3.1
amqp==5.0.9
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.3
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.1
Cython==0.29.28
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.2
django-celery-beat==2.2.1
django-cors-headers==3.11.0
django-rest-framework==0.1.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
GDAL==3.4.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.3
msgpack==1.0.3
numpy==1.22.2
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.1.0
txaio==22.2.1
typing_extensions==4.1.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zope.interface==5.4.0
